make Hanami::Validations and Hanami::Validations::Form classes instead of mixin modules
