+++ spec/unit/hanami/validations/base_validations_spec.rb
@@ -1,9 +1,7 @@
 RSpec.describe Hanami::Validations do
   describe 'base validations' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+      @validator = Class.new(Hanami::Validations) do
         def self._base_rules
           lambda do
             optional(:_csrf_token).filled(:str?)
