+++ spec/unit/hanami/validations/form/rules_spec.rb
@@ -3,9 +3,7 @@ require 'uri'
 RSpec.describe Hanami::Validations::Form do
   describe 'rules' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations::Form
-
+      @validator = Class.new(Hanami::Validations::Form) do
         validations do
           required(:type).filled(:int?, included_in?: [1, 2, 3])
