+++ spec/unit/hanami/validations/predicates/form/custom_spec.rb
@@ -3,9 +3,7 @@ RSpec.describe 'Predicates: custom' do
 
   describe 'with custom predicate' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations::Form
-
+      @validator = Class.new(Hanami::Validations::Form) do
         def self.name
           'Validator'
         end
@@ -43,9 +41,7 @@ RSpec.describe 'Predicates: custom' do
 
   describe 'with custom predicates as module' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations::Form
-
+      @validator = Class.new(Hanami::Validations::Form) do
         def self.name
           'Validator'
         end
@@ -86,9 +82,7 @@ RSpec.describe 'Predicates: custom' do
 
   describe 'with custom predicate within predicates block' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations::Form
-
+      @validator = Class.new(Hanami::Validations::Form) do
         def self.name
           'Validator'
         end
@@ -122,9 +116,7 @@ RSpec.describe 'Predicates: custom' do
 
   describe 'with custom predicate with predicate macro' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations::Form
-
+      @validator = Class.new(Hanami::Validations::Form) do
         def self.name
           'Validator'
         end
@@ -165,9 +157,7 @@ RSpec.describe 'Predicates: custom' do
   describe 'without custom predicate' do
     it 'raises error if try to use an unknown predicate' do
       expect do
-        Class.new do
-          include Hanami::Validations::Form
-
+        Class.new(Hanami::Validations::Form) do
           def self.name
             'Validator'
           end
@@ -182,9 +172,7 @@ RSpec.describe 'Predicates: custom' do
 
   describe 'with nested validations' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations::Form
-
+      @validator = Class.new(Hanami::Validations::Form) do
         def self.name
           'Validator'
         end
