+++ spec/unit/hanami/validations/predicates/form/min_size_spec.rb
@@ -3,9 +3,7 @@ RSpec.describe 'Predicates: Min Size' do
 
   describe 'with required' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations::Form
-
+      @validator = Class.new(Hanami::Validations::Form) do
         validations do
           required(:foo) { min_size?(3) }
         end
@@ -55,9 +53,7 @@ RSpec.describe 'Predicates: Min Size' do
 
   describe 'with optional' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations::Form
-
+      @validator = Class.new(Hanami::Validations::Form) do
         validations do
           optional(:foo) { min_size?(3) }
         end
@@ -109,9 +105,7 @@ RSpec.describe 'Predicates: Min Size' do
     describe 'with required' do
       describe 'with value' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations::Form
-
+          @validator = Class.new(Hanami::Validations::Form) do
             validations do
               required(:foo).value(min_size?: 3)
             end
@@ -161,9 +155,7 @@ RSpec.describe 'Predicates: Min Size' do
 
       describe 'with filled' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations::Form
-
+          @validator = Class.new(Hanami::Validations::Form) do
             validations do
               required(:foo).filled(min_size?: 3)
             end
@@ -213,9 +205,7 @@ RSpec.describe 'Predicates: Min Size' do
 
       describe 'with maybe' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations::Form
-
+          @validator = Class.new(Hanami::Validations::Form) do
             validations do
               required(:foo).maybe(min_size?: 3)
             end
@@ -267,9 +257,7 @@ RSpec.describe 'Predicates: Min Size' do
     describe 'with optional' do
       describe 'with value' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations::Form
-
+          @validator = Class.new(Hanami::Validations::Form) do
             validations do
               optional(:foo).value(min_size?: 3)
             end
@@ -319,9 +307,7 @@ RSpec.describe 'Predicates: Min Size' do
 
       describe 'with filled' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations::Form
-
+          @validator = Class.new(Hanami::Validations::Form) do
             validations do
               optional(:foo).filled(min_size?: 3)
             end
@@ -371,9 +357,7 @@ RSpec.describe 'Predicates: Min Size' do
 
       describe 'with maybe' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations::Form
-
+          @validator = Class.new(Hanami::Validations::Form) do
             validations do
               optional(:foo).maybe(min_size?: 3)
             end
