+++ spec/unit/hanami/validations/predicates/schema/array_spec.rb
@@ -3,9 +3,7 @@ RSpec.describe 'Predicates: Array' do
 
   describe 'with required' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+      @validator = Class.new(Hanami::Validations) do
         validations do
           required(:foo) { array? { each { int? } } }
         end
@@ -82,9 +80,7 @@ RSpec.describe 'Predicates: Array' do
 
   describe 'with optional' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+      @validator = Class.new(Hanami::Validations) do
         validations do
           optional(:foo) { included_in?([1, 3, 5]) }
         end
@@ -143,9 +139,7 @@ RSpec.describe 'Predicates: Array' do
   describe 'as macro' do
     describe 'with required' do
       before do
-        @validator = Class.new do
-          include Hanami::Validations
-
+        @validator = Class.new(Hanami::Validations) do
           validations do
             required(:foo).each(:int?)
           end
@@ -195,9 +189,7 @@ RSpec.describe 'Predicates: Array' do
 
     describe 'with optional' do
       before do
-        @validator = Class.new do
-          include Hanami::Validations
-
+        @validator = Class.new(Hanami::Validations) do
           validations do
             optional(:foo).each(:int?)
           end
