+++ spec/unit/hanami/validations/predicates/schema/empty_spec.rb
@@ -3,9 +3,7 @@ RSpec.describe 'Predicates: Empty' do
 
   describe 'with required' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+      @validator = Class.new(Hanami::Validations) do
         validations do
           required(:foo) { empty? }
         end
@@ -63,9 +61,7 @@ RSpec.describe 'Predicates: Empty' do
 
   describe 'with optional' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+      @validator = Class.new(Hanami::Validations) do
         validations do
           optional(:foo) { empty? }
         end
@@ -125,9 +121,7 @@ RSpec.describe 'Predicates: Empty' do
     describe 'with required' do
       describe 'with value' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               required(:foo).value(:empty?)
             end
@@ -207,9 +201,7 @@ RSpec.describe 'Predicates: Empty' do
     describe 'with optional' do
       describe 'with value' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               optional(:foo).value(:empty?)
             end
