+++ spec/unit/hanami/validations/predicates/schema/excluded_from_spec.rb
@@ -3,9 +3,7 @@ RSpec.describe 'Predicates: Excluded From' do
 
   describe 'with required' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+      @validator = Class.new(Hanami::Validations) do
         validations do
           required(:foo) { excluded_from?([1, 3, 5]) }
         end
@@ -63,9 +61,7 @@ RSpec.describe 'Predicates: Excluded From' do
 
   describe 'with optional' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+      @validator = Class.new(Hanami::Validations) do
         validations do
           optional(:foo) { excluded_from?([1, 3, 5]) }
         end
@@ -125,9 +121,7 @@ RSpec.describe 'Predicates: Excluded From' do
     describe 'with required' do
       describe 'with value' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               required(:foo).value(excluded_from?: [1, 3, 5])
             end
@@ -185,9 +179,7 @@ RSpec.describe 'Predicates: Excluded From' do
 
       describe 'with filled' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               required(:foo).filled(excluded_from?: [1, 3, 5])
             end
@@ -245,9 +237,7 @@ RSpec.describe 'Predicates: Excluded From' do
 
       describe 'with maybe' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               required(:foo).maybe(excluded_from?: [1, 3, 5])
             end
@@ -307,9 +297,7 @@ RSpec.describe 'Predicates: Excluded From' do
     describe 'with optional' do
       describe 'with value' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               optional(:foo).value(excluded_from?: [1, 3, 5])
             end
@@ -367,9 +355,7 @@ RSpec.describe 'Predicates: Excluded From' do
 
       describe 'with filled' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               optional(:foo).filled(excluded_from?: [1, 3, 5])
             end
@@ -427,9 +413,7 @@ RSpec.describe 'Predicates: Excluded From' do
 
       describe 'with maybe' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               optional(:foo).maybe(excluded_from?: [1, 3, 5])
             end
