+++ spec/unit/hanami/validations/predicates/schema/type_spec.rb
@@ -3,9 +3,7 @@ RSpec.describe 'Predicates: Type' do
 
   describe 'with required' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+      @validator = Class.new(Hanami::Validations) do
         validations do
           required(:foo) { type?(Integer) }
         end
@@ -55,9 +53,7 @@ RSpec.describe 'Predicates: Type' do
 
   describe 'with optional' do
     before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+      @validator = Class.new(Hanami::Validations) do
         validations do
           optional(:foo) { type?(Integer) }
         end
@@ -109,9 +105,7 @@ RSpec.describe 'Predicates: Type' do
     describe 'with required' do
       describe 'with value' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               required(:foo).value(type?: Integer)
             end
@@ -161,9 +155,7 @@ RSpec.describe 'Predicates: Type' do
 
       describe 'with filled' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               required(:foo).filled(type?: Integer)
             end
@@ -213,9 +205,7 @@ RSpec.describe 'Predicates: Type' do
 
       describe 'with maybe' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               required(:foo).maybe(type?: Integer)
             end
@@ -267,9 +257,7 @@ RSpec.describe 'Predicates: Type' do
     describe 'with optional' do
       describe 'with value' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               optional(:foo).value(type?: Integer)
             end
@@ -319,9 +307,7 @@ RSpec.describe 'Predicates: Type' do
 
       describe 'with filled' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               optional(:foo).filled(type?: Integer)
             end
@@ -371,9 +357,7 @@ RSpec.describe 'Predicates: Type' do
 
       describe 'with maybe' do
         before do
-          @validator = Class.new do
-            include Hanami::Validations
-
+          @validator = Class.new(Hanami::Validations) do
             validations do
               optional(:foo).maybe(type?: Integer)
             end
